Seperated to save and load methods into SavePlayerLocation class
