+++ Assets/02.Scripts/FSM/State/StateController/PlayerStateController.cs
@@ -262,24 +262,6 @@ public class PlayerStateController : MonoBehaviour, IStateController
         }
     }
 
-    public void SavePlayer()
-    {
-        SaveSystem.SavePlayer(this);
-    }
-
-    public void LoadPlayer()
-    {
-        PlayerStats data = SaveSystem.LoadPlayer();
-
-        Vector3 position;
-        position.x = data.position[0];
-        position.y = data.position[1];
-        position.z = data.position[2];
-
-        playerTranslate.Character.position = position;
-        Transform.position = position;
-    }
-
 }
