+++ Assets/02.Scripts/FSM/Stats/SavePlayerLocation.cs
@@ -1,0 +1,38 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SavePlayerLocation : MonoBehaviour
+{
+
+    //// Start is called before the first frame update
+    //void Start()
+    //{
+
+    //}
+
+    //// Update is called once per frame
+    //void Update()
+    //{
+
+    //}
+
+    public void SavePlayer()
+    {
+        Debug.Log("Button pressed");
+        SaveSystem.SavePlayer(this);
+    }
+
+    public void LoadPlayer()
+    {
+        PlayerStats data = SaveSystem.LoadPlayer();
+
+        Vector3 position;
+        position.x = data.position[0];
+        position.y = data.position[1];
+        position.z = data.position[2];
+
+        //playerTranslate.Character.position = position;
+        transform.position = position;
+    }
+}
