+++ Assets/02.Scripts/FSM/Stats/SavePlayerLocation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c822d44c703ee80489952020d1d548e0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
